fonctionnalites des commentaire et ajustement des publications

+++ src/main/java/com/frank/apisocialnetwork/repository/TokenRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.Optional;
 
 
-public interface TokenRepository extends JpaRepository<Token, Integer> {
+public interface TokenRepository extends JpaRepository<Token, Long> {
 
     List<Token> findByRefreshToken(String bearer);
 }
